Add files via upload
It is a project which I completed under an internship program in Skillvertex. Here datasets were given to us and we tested the data and trained it to find the best fit model for the dataset. The best model which I found for the data given is Random Forest Classifier.